requirements.txt
langchain_core
langchain_community
dotenv
streamlit
requests
google-search-results